Add support for activating the applications we launch
Adds the necessary bits for implementing xdg_activation_v1 for
applications launched with global shortcuts.

It adds a KGlobalAccelInterfaceV3 that the compositor should implement
and will be used to request a token. This token is forwarded to the
processes we launch either through dbus activation or as the environment
variable.