For D-Bus de/marshalling types, use int instead of KGlobalAccel::MatchType
And cast to/from int in the two places where that is needed. Otherwise
the globalShortcutsByKey() method is silently skipped, and isn't visible on
the org.kde.KGlobalAccel D-Bus interface.

It turns out that to test changes made in kglobalaccel d-bus interface, you
need to kill the system /usr/bin/kglobalaccel process, and run the one from
the build-dir, otherwise the session d-bus won't pick up your changes.

To test:
- List all methods on the interface:
  `qdbus org.kde.kglobalaccel /kglobalaccel`
- The followin messages are printed in the system log/journal:
Skipped method "globalShortcutsByKey" : Unregistered input type in parameter list: KGlobalAccel::MatchType
Skipped method "globalShortcutsByKey" : Unregistered input type in parameter list: KGlobalAccel::MatchType

To see the behaviour mentioned in the bug report:
- Create a global shortcut of some kind (e.g. F4 to start some app)
- open the shortcut editor in e.g. Dolphin, and try to assign that same
  shortcut to some other action the conflict shortcut message dialog should
  show a message similar to the one in the attached screenshot at:
  https://bugs.kde.org/show_bug.cgi?id=454704#c0
The list of GlobalShortcutInfo returned by globalShortcutsByKey() is empty
because the method couldn't be found on the d-bus interface.

BUG: 454704
FIXED-IN: 5.95