Fix connection lifetime issue
The lack of receiver may cause undefined behavior / segmentation fault
crash if, for example, a KActionCollection is stored as a
Q_GLOBAL_STATIC too and coincidentally gets destructed after the
KGlobalAccel singleton instance.

Test case: Run `plasmoidviewer -a org.kde.plasma.clipboard`. With this
patch is still crashes on exit, but now in a different place.

CCBUG:454854
CCBUG:485386